[INJIWEB-1704] - Rename ClientValidationUtils to UrlParameterUtils for better clarity - Simplify verifier validation logic by removing unnecessary exception handling - Add proper URI syntax error handling in WalletPresentationsController - Update vcverifier-jar dependency to 1.5.0-SNAPSHOT - Enhance test coverage with additional edge cases for URL parameter extraction - Fix test assertion for verifier pre-registration validation - Add new API response example for invalid URI syntax errors
